+++ .eslintrc.json
@@ -31,6 +31,7 @@
         "plugin:prettier/recommended"
       ],
       "rules": {
+        "@typescript-eslint/no-explicit-any": "off",
         "@typescript-eslint/no-shadow": ["error", { "ignoreTypeValueShadow": true }],
         "@typescript-eslint/no-use-before-define": 0,
         "class-methods-use-this": 0,
