+++ package.json
@@ -143,7 +143,7 @@
     "@types/express": "^4.17.21",
     "@types/express-session": "^1.18.0",
     "@types/http-errors": "^2.0.4",
-    "@types/jest": "^29.5.12",
+    "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/node": "^20.11.4",
     "@types/nunjucks": "^3.2.6",
@@ -181,7 +181,7 @@
     "prettier-plugin-jinja-template": "^1.5.0",
     "sass": "^1.77.8",
     "supertest": "^7.0.0",
-    "ts-jest": "^29.2.5",
+    "ts-jest": "^29.3.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   },
