+++ server/data/probationOffenderSearch/probationOffenderSearchApiClient.ts
@@ -2,7 +2,7 @@ import Agent, { HttpsAgent } from 'agentkeepalive'
 import superagent from 'superagent'
 import { config } from '../../config'
 import { getApiClientToken } from '../../authentication/clientCredentials'
-import getSanitisedError from '../../sanitisedError'
+import { getSanitisedError } from '../../getSanitisedError'
 import logger from '../../../log'
 import { ProbationOffenderSearchOffenderDto } from './probationOffenderSearchOffender.dto'
 import { User } from '../user'
