+++ server/data/riskProfilerClientBuilder.js
@@ -4,7 +4,7 @@ const { HttpsAgent } = require('agentkeepalive')
 const logger = require('../../log')
 const { config } = require('../config')
 const { getApiClientToken } = require('../authentication/clientCredentials')
-const getSanitisedError = require('../sanitisedError')
+const { getSanitisedError } = require('../getSanitisedError')
 
 const timeoutSpec = {
   response: config.apis.riskProfiler.timeout.response,
