+++ server/data/risksAndNeeds/risksAndNeedsApi.ts
@@ -4,7 +4,7 @@ import redis from 'redis'
 import { promisify } from 'util'
 import { config } from '../../config'
 import { getApiClientToken } from '../../authentication/clientCredentials'
-import getSanitisedError from '../../sanitisedError'
+import { getSanitisedError } from '../../getSanitisedError'
 import logger from '../../../log'
 import { User } from '../user'
 import { RiskSummaryDto } from './riskSummary.dto'
